Add files via upload
The goal of the following code is to search for a specific date and time values (introduced by the user) in a CSV file logged by different sensors. The CSV file includes reading values such as pressure, temperature, altitude, depth, time, and location among others. The data come from attached sensors to a sonde (electronic board).

The code checks in the CSV file if the date and time exist in the same reading line of the CSV file (no matter the order of the time and date columns in the CSV file is) and if so it prints out the reading values of all the sensors associated/dated with that date and time.